[Fix] Fix tests for testing notification preferences for disable orgs #367
Changed location of the test to test_api.py and removed single use function for efficiency.

Fixes #367
